Added my Test3.html backup and DarkBasicPro proj.
The Test3.html file is just a collection of stuff I thought was useful but seldom refer back to anymore. So I thought I'd share it anyway. Maybe someone else may find it of interest. The DbProSlideDemo is a small 3D "world" I made back in the day in DarkBasicPro and is so far the only one that big that I ever made. I mostly built it trying to teach myself 3D Matrix math.