Encryption tool I wrote in Java - Lurch
Including this here as well. I wrote an encryption tool in Java when I was stationed in Iraq in the US Army. 
I submitted a copy of it to the NSA already and it's nothing so fancy as to worry about. It handles pretty
big file directories very well but is a little wonky to use so read the readme file for sure. (Also as my real name
is included in that, don't be surprised - Warmag and Draxxar were my EverQuest characters by the way so I
go by those mostly when I'm doing online stuff - except for here now).